Balayage complet: 208 backtests indiciels re-executes, 5 PASS de plus tombent
42 fichiers de resultat modifies, 5 PASS->FAIL, 0 FAIL->PASS, 37 scores
modifies a verdict constant. Bilan consolide #375->#377: 12 PASS tombes,
0 gagne, aucun verdict ne s'ameliore.

4 echecs rencontres et resolus: 2 incompatibilites pandas>=3 (sans rapport
avec le bug), 1 script exigeant un argument de marche, 1 dependant du premier.

Reste suspects: les 31 scripts de portefeuille au niveau titre, porteurs d'un
defaut distinct (moyenne ponderee de rendements log) non corrige.

+++ finance/trading/scripts/nonml_multimarket_breadth_vol_targeting_overlay_backtest.py
@@ -88,7 +88,9 @@ def main():
     bh_full = np.log(close[1:] / close[:-1])
 
     breadth = compute_multimarket_breadth_series(dates_idx)
-    breadth_raw = breadth.values
+    # astype(float) : sous pandas >= 3 la serie peut ressortir en dtype object,
+    # que np.isnan refuse. Conversion explicite, aucun changement de valeur.
+    breadth_raw = breadth.values.astype(float)
     gate_aligned = np.where(np.isnan(breadth_raw), False, breadth_raw >= BREADTH_THRESHOLD)
 
     pos_full = combined_position(bh_full, gate_aligned)
